1.3 add view

+++ src/main/java/com/example/votingsystem/repository/JpaDishRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JpaDishRepo extends JpaRepository<Dish, Long> {
+public interface JpaDishRepo extends JpaRepository<Dish, Integer> {
 }
